google_add/requirements.txt
# Agentic AI Google Ads Generator - Requirements

# Core dependencies
fastapi>=0.128.5
uvicorn>=0.40.0
pydantic>=2.12.5
python-multipart>=0.0.22

# AI/ML dependencies
aiohttp>=3.8.0
Pillow>=10.0.0

# Google Ads API
google-ads>=22.0.0

# Already in your project
python-dotenv>=1.2.1
requests>=2.28
numpy>=2.4.2
Flask>=2.0
click>=8.0
